actions: Split out zerver.actions.muted_users.
